Remove windows support Only one dev uses it, and it's not even supported officially. It would be easier long-term to simply kill off this feature before v3 is actually live.
