+++ .github/workflows/build.yml
@@ -112,142 +112,6 @@ jobs:
           if-no-files-found: error
           retention-days: 1
 
-  build-windows-amd64-binaries:
-    name: Build Binaries - Windows AMD64
-    runs-on: ubuntu-latest
-    needs: [ 'build-shared' ]
-
-    timeout-minutes: 10
-
-    steps:
-      - name: Get the version
-        run: echo "VERSION=0.0.${GITHUB_RUN_NUMBER}" >> $GITHUB_ENV
-
-      - name: Get the version
-        if: startsWith(github.ref, 'refs/tags/v')
-        run: echo "VERSION=${GITHUB_REF/refs\/tags\/v/}" >> $GITHUB_ENV
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: pufferpanel_build_artifact
-
-      - name: Set up Go
-        uses: actions/setup-go@v5
-        with:
-          go-version-file: 'go.mod'
-
-      - name: Download APT packages
-        run: |
-          sudo apt-get update
-          sudo apt-get install --download-only -y gcc-mingw-w64
-
-      - name: APT Install
-        run: |
-          sudo apt-get install -y gcc-mingw-w64
-
-      - name: Build Windows
-        run: |
-          go build -o build-artifacts/pufferpanel.exe -ldflags "-X 'github.com/pufferpanel/pufferpanel/v3.Hash=${{ github.sha }}' -X 'github.com/pufferpanel/pufferpanel/v3.Version=${{ env.VERSION }}'" -v github.com/pufferpanel/pufferpanel/v3/cmd
-        env:
-          GOOS: windows
-          GOARCH: amd64
-          CC: x86_64-w64-mingw32-gcc
-          CXX: x86_64-w64-mingw32-g++
-          CGO_ENABLED: 1
-
-      - name: Publish Artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: pufferpanel_${{ env.VERSION }}_windows_amd64
-          path: build-artifacts/
-
-  build-windows-installer:
-    name: Build Windows Installer
-    runs-on: windows-latest
-    needs: [ 'build-windows-amd64-binaries' ]
-
-    env:
-      DOTNET_CHANNEL: '8.0'
-      WIX_VERSION: '4.0.4'
-
-    timeout-minutes: 10
-
-    steps:
-      - name: Get the version
-        run: |
-          $VERSION="0.0.${{ github.run_number }}"
-          if ("${{ github.ref }}".startsWith("refs/tags/v")) {
-            $VERSION="${{ github.ref }}".substring('refs/tags/v'.length)
-          }
-          echo VERSION=$VERSION | Out-File -FilePath $Env:GITHUB_ENV -Encoding utf-8 -Append
-
-      #      - name: Restore DOTNET BIN
-      #        uses: actions/cache@v3
-      #        id: dotnet-cache
-      #        with:
-      #          path: |
-      #            ~\dotnet
-      #          key: ${{ runner.os }}-dotnet-${{ env.DOTNET_CHANNEL }}
-      #          restore-keys: |
-      #            ${{ runner.os }}-dotnet-
-      #
-      #      - name: Restore Wix cache
-      #        uses: actions/cache@v3
-      #        id: wix-cache
-      #        with:
-      #          path: |
-      #            ~\.wix
-      #            ~\.dotnet
-      #          key: ${{ runner.os }}-wix-${{ env.DOTNET_CHANNEL }}
-      #          restore-keys: |
-      #            ${{ runner.os }}-wix-
-
-      - name: Prepare .NET
-        #if: steps.dotnet-cache.outputs.cache-hit != 'true'
-        run: |
-          Invoke-WebRequest "https://dot.net/v1/dotnet-install.ps1" -OutFile dotnet-install.ps1
-          .\dotnet-install.ps1 -Channel ${{ env.DOTNET_CHANNEL }} -InstallDir dotnet
-          Remove-Item dotnet-install.ps1
-
-      - name: Set Path
-        run: |
-          Add-Content $env:GITHUB_PATH "$pwd\dotnet"
-          Add-Content $env:GITHUB_PATH "$pwd\.dotnet\tools"
-          Add-Content $env:GITHUB_ENV "DOTNET_ROOT=$pwd\dotnet"
-
-      - name: Install WIX
-        #if: steps.wix-cache.outputs.cache-hit != 'true'
-        run: |
-          dotnet tool install --tool-path .dotnet\tools wix --version ${{ env.WIX_VERSION }}
-          wix extension add WixToolset.Firewall.wixext/${{ env.WIX_VERSION }}
-          wix extension add WixToolset.UI.wixext/${{ env.WIX_VERSION }}
-          wix extension add WixToolset.Util.wixext/${{ env.WIX_VERSION }}
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: pufferpanel_${{ env.VERSION }}_windows_amd64
-          path: binaries
-
-      - uses: actions/download-artifact@v4
-        with:
-          name: pufferpanel_build_artifact
-          path: artifacts
-
-      - name: Generate MSI
-        run: |
-          Copy-Item -Recurse -Path binaries\* -Destination .
-          Copy-Item -Path artifacts\client\frontend\public\favicon.ico -Destination .
-          Copy-Item -Path artifacts\LICENSE.rtf -Destination .
-          Copy-Item -Path artifacts\installer.wxs -Destination .
-          set password $((New-Guid).ToString())
-          wix build -o pufferpanel.msi -arch x64 -d ProductVersion="${{ env.VERSION }}" -d Password="$password" -ext WixToolset.Firewall.wixext -ext WixToolset.UI.wixext -ext WixToolset.Util.wixext installer.wxs
-
-      - name: Publish Artifacts
-        uses: actions/upload-artifact@v4
-        with:
-          name: pufferpanel_${{ env.VERSION }}_installer.msi
-          path: pufferpanel.msi
-
   build-linux-amd64-binaries:
     name: Build Binaries - Linux AMD64
     runs-on: ubuntu-latest
@@ -500,7 +364,7 @@ jobs:
     runs-on: ubuntu-latest
     name: "Cleanup"
 
-    needs: [ build-shared, build-docker-image-base, build-windows-amd64-binaries, build-windows-installer, build-linux-amd64-binaries, build-linux-arm64-binaries, release ]
+    needs: [ build-shared, build-docker-image-base, build-linux-amd64-binaries, build-linux-arm64-binaries, release ]
 
     continue-on-error: true
 
@@ -596,8 +460,6 @@ jobs:
     runs-on: ubuntu-latest
     needs:
       - build-docker-image-base
-      - build-windows-amd64-binaries
-      - build-windows-installer
       - build-linux-amd64-binaries
       - build-linux-arm64-binaries
     permissions:
@@ -643,4 +505,3 @@ jobs:
             pufferpanel_*_windows_amd64
             */*.rpm
             */*.deb
-            */*.msi
