+++ client/package.json
@@ -1,8 +1,12 @@
 {
   "private": true,
-  "workspaces": ["api", "frontend"],
+  "workspaces": [
+    "api",
+    "frontend"
+  ],
   "scripts": {
     "build": "yarn workspaces run build",
     "dev": "yarn workspaces run dev"
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
