+++ config/config.go
@@ -5,7 +5,6 @@ import (
 	"github.com/spf13/viper"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 )
 
@@ -20,16 +19,9 @@ func LoadConfigFile(configFile string) error {
 		var exists bool
 		configFile, exists = os.LookupEnv("PUFFER_CONFIG")
 		if !exists || configFile == "" {
-			if runtime.GOOS == "windows" {
-				//well, check for the program files path....
-				if _, err := os.Lstat("C:\\Program Files\\PufferPanel\\config.json"); err == nil {
-					configFile = "C:\\Program Files\\PufferPanel\\config.json"
-				}
-			} else {
-				//well, check for the /etc path
-				if _, err := os.Lstat("/etc/pufferpanel/config.json"); err == nil {
-					configFile = "/etc/pufferpanel/config.json"
-				}
+			//well, check for the /etc path
+			if _, err := os.Lstat("/etc/pufferpanel/config.json"); err == nil {
+				configFile = "/etc/pufferpanel/config.json"
 			}
 			//we got nothing
 			if configFile == "" {
