--- groups/groups_windows.go
@@ -1,7 +1,0 @@
-//go:build !linux
-
-package groups
-
-func IsUserIn(groups ...string) bool {
-	return true
-}
