--- logging/logger_noservice.go
@@ -1,11 +1,0 @@
-//go:build !windows
-
-package logging
-
-import (
-	"io"
-)
-
-func CreateServiceLogger(string) io.Writer {
-	return nil
-}
