+++ logging/rotator.go
@@ -2,7 +2,11 @@ package logging
 
 import (
 	"io"
+	"os"
+	"os/signal"
+	"path"
 	"sync"
+	"syscall"
 )
 
 type Rotator struct {
@@ -33,4 +37,22 @@ func (r *Rotator) Rotate(newBackend io.WriteCloser) {
 	oldBacker := r.backer
 	r.backer = newBackend
 	_ = oldBacker.Close()
+}
+
+func (r *Rotator) StartRotation(dir string) {
+	go func(directory string) {
+		sig := make(chan os.Signal, 1)
+		for {
+			signal.Notify(sig, syscall.SIGUSR1)
+
+			<-sig
+
+			newFile, err := os.OpenFile(path.Join(directory, "pufferpanel.log"), os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0644)
+			if err != nil {
+				panic(err)
+			}
+
+			r.Rotate(newFile)
+		}
+	}(dir)
 }
