--- logging/rotator_windows.go
@@ -1,4 +1,0 @@
-package logging
-
-func (r *Rotator) StartRotation(dir string) {
-}
