+++ operations/curseforge/const.go
@@ -5,6 +5,6 @@ const ImprovedFabricInstallerUrl = "https://meta.fabricmc.net/v2/versions/loader
 const ForgeInstallerUrl = "https://maven.minecraftforge.net/net/minecraftforge/forge/${mcVersion}-${version}/forge-${mcVersion}-${version}-installer.jar"
 const ForgeInstallerName = "forge-${mcVersion}-${version}-installer.jar"
 const NeoforgeInstallerUrl = "https://maven.minecraftforge.net/net/minecraftforge/forge/${mcVersion}-${version}/forge-${mcVersion}-${version}-installer.jar"
-const NeoforgeInstallerName = "neoforge-${version}-installer.jar"
+const NeoforgeInstallerName = "neoforgedl-${version}-installer.jar"
 const NeoForgeServerStarterVersion = "0.1.25"
 const NeoForgeServerStarter = "https://github.com/neoforged/ServerStarterJar/releases/download/" + NeoForgeServerStarterVersion + "/server.jar"
