+++ servers/tty/factory.go
@@ -1,5 +1,3 @@
-//go:build !windows
-
 package tty
 
 import (
