--- utils/kernel_windows.go
@@ -1,5 +1,0 @@
-package utils
-
-func testOpenat2() {
-	useOpenat2 = false
-}
