+++ web/tests/server_test.go
@@ -25,7 +25,6 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"testing"
 	"time"
@@ -722,11 +721,6 @@ func TestServers(t *testing.T) {
 				})
 
 				t.Run("DeleteFileWithURIEncoding", func(t *testing.T) {
-					if runtime.GOOS == "windows" {
-						t.Skipf("Windows doesn't support file with URI encoding")
-						return
-					}
-
 					filename := "file.delete.test?id=12345"
 
 					fileLocation := filepath.Join(serverDir, filename)
